skia: Optimize skia BLIT routine using neon instructions
Use neon instructions to optimize BLIT routine which operates on
32-bit source ARGB pixel and 16-bit destination RGB pixel. Here,
we process 16-pixels in an iteration. We fall back to regular (32-bit)
arm instructions when number of pixels is less than 16 and/or the
architecture doesn't support SIMD/NEON instructions. When alpha value
is 255, this version of BLIT routine performs 35% better in comparison
to S32A_D565_Opaque_neon routine.